Struct hamt::arc::HamtSetArc

pub struct HamtSetArc<K> {
    // some fields omitted
}

Methods

impl<K> HamtSetArc<K> where K: Hash + Eq + Clone

fn new() -> Self

Returns an empty set.

fn len(&self) -> usize

Returns how many items are in the set.

fn is_empty(&self) -> bool

Returns true if the set is empty, false otherwise.

fn insert<Q: ?Sized>(&self, k: &Q) -> Self where K: Borrow<Q>, Q: Hash + Eq + ToOwned<Owned=K>

Inserts an item into the set.

fn contains<Q: ?Sized>(&self, k: &Q) -> bool where K: Borrow<Q>, Q: Hash + Eq

Returns true if the set contains the given item, false otherwise.

fn remove<Q: ?Sized>(&self, k: &Q) -> Self where K: Borrow<Q>, Q: Hash + Eq

Removes the given item from the set.

fn iter(&self) -> Keys<K, ()>

Iterates over each item in the set in an unspecified order.